+++ components/icon/src/Icon/hooks/useIconProps.tsx
@@ -1,8 +1,8 @@
 import { type KeyboardEvent, type MouseEvent, type UIEvent, useCallback, useMemo, useRef } from "react";
 import cx from "classnames";
-import { noop as NOOP } from "es-toolkit/compat";
 import { useEventListener, useKeyEvent, keyCodes } from "@vibe/shared";
 import useIconScreenReaderAccessProps from "./useIconScreenReaderAccessProps";
+import { noop as NOOP } from "es-toolkit";
 import styles from "../Icon.module.scss";
 
 const KEYS = [keyCodes.ENTER, keyCodes.SPACE];