+++ packages/core/src/components/Combobox/components/ComboboxConstants.ts
@@ -97,6 +97,11 @@ export interface IComboboxOption {
    * The tooltip content displayed on hover.
    */
   tooltipContent?: string;
+  /**
+   * Props for the Tooltip component.
+   * The "Omit" is here on purpose to avoid the content prop, in case tooltipProps would be expanded in the future.
+   */
+  tooltipProps?: Pick<Omit<TooltipProps, "content">, "position">;
 }
 
 export interface IComboboxItem {