+++ packages/core/src/storybook/stand-alone-documentaion/ux-writing-handbook/values/rulesUXWriting.tsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { bottomLine, getFull, userPointOfView, oneCta, specificCta, cutTheFluff } from "../assets";
+import { Value } from "./values";
+import styles from "./valuesUXWriting.module.scss";
+
+export const RulesUXWriting = () => (
+  <div className={styles.principles}>
+    <Value
+      imgSrc={bottomLine}
+      title="Bottom line first"
+      description="Start every message with the most important information."
+    />
+    <Value
+      imgSrc={getFull}
+      title="Give full context"
+      description="Provide enough context so users know exactly where they are."
+    />
+    <Value
+      imgSrc={userPointOfView}
+      title="Focus on the user's point of view"
+      description="Focus on what the user wants to do and what they stand to gain, not on your own perspective."
+    />
+    <Value
+      imgSrc={oneCta}
+      title="One CTA"
+      description="Present only one CTA at a time. It makes decisions easier and faster."
+    />
+    <Value
+      imgSrc={specificCta}
+      title="Use specific CTAs"
+      description="Ensure that CTAs respond directly to the header."
+    />
+    <Value imgSrc={cutTheFluff} title="Cut the fluff" description="Write briefly and to the point." />
+  </div>
+);